CH340G FTDI Programmer V2.0
The board is ready to order, but I can`t order it rigth now, so I will have to wait about 2 months to finish this project permanently.

So the last version had the D+ and D- reversed, IDK how I did that, and more than that, I inversed the TX and RX pins and not realy. Iconnected the atmega tx to an conexion named rx, expecting ro connect RX to that pin, and same with rx. The same I done with the FTDI, and by doing that, I double reversed the tx and rx wires, so it`s like I didn`t reversed them at all.
In all that tests I saw that the 100nF capacitor at the DTR is not necesarry, and it could be just a wire in it`s place, and more than that, could be added some design changes... round corners, holes for a case, tx and rx yellow led`s, and stuff like this.